pyproject.toml
[tool.poetry]
name = "todo-fastapi"
version = "0.1.0"
description = ""
authors = ["Henish Patel <200120107182@git.org.in>"]
readme = "README.md"
packages = [{include = "todo_fastapi", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
alembic = "^1.13.1"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
mysqlclient = "^2.2.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
